requirements.txt
spidev
mfrc522
mysql.connector
datetime
